requirements.txt
streamlit
opencv-python
numpy
Pillow
tensorflow
python-dotenv
langchain
langchain-groq
langchain-core
langchain-community
langchain-google-genai
faiss-cpu
PyPDF2
pypdf
scikit-learn
google-generativeai